+++ backend_src/backend/pages/templates/pages/main.html
@@ -9,6 +9,6 @@
     <div id="root"></div>
     <!-- Vite 開発サーバーからのファイルを読み込み -->
     <script src="http://localhost:5173/@vite/client" type="module"></script>
-    <script src="http://localhost:5173/src/topPage.tsx" type="module"></script>
+    <script src="http://localhost:5173/src/main.tsx" type="module"></script>
 </body>
 </html>
