+++ backend_src/backend/pages/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
-from .views import PageView
+from .views import MainEntryView
+
 
 urlpatterns = [
-    path('pages/', PageView.as_view(), name='page-list'),
+    # path('pages/', MainEntryView.as_view(), name='page-list'),  # 必要に応じて名前を変更
 ]
