+++ frontend_src/node/frontend/index.html
@@ -8,6 +8,6 @@
   <body>
     <!-- Reactアプリをマウントするためのdiv -->
     <div id="root"></div>
-    <script type="module" src="/src/topPage.tsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
